refactor: move note definitions into separate file
