+++ src/components/Score.test.tsx
@@ -2,7 +2,8 @@ import { render, RenderResult, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 import { FormFactor } from '../App';
-import { getGeneratedNoteGroups, NoteGroupType } from '../modules/note';
+import { getGeneratedNoteGroups } from '../modules/note';
+import { NoteGroupType } from '../modules/note-definition';
 import { getTimeSignature, TimeSignatureType } from '../modules/time-signature';
 import * as vex from '../modules/vex';
 import Score, { scoreDimensionConfig } from './Score';
